requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
sentence-transformers
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-chroma==0.1.4
pdfminer.six
python-docx
langchain-groq
pydantic
python-multipart
python-jose
passlib[bcrypt]
requests
pypdf
redis
sqlalchemy
pymongo
langchain-chroma
